requirements.txt
# 核心Web框架
fastapi==0.104.1
uvicorn[standard]==0.24.0

# WebSocket支持 
websockets==12.0

# Redis客户端
redis==5.0.1

# MQTT客户端
paho-mqtt==1.6.1

# HTTP客户端
httpx==0.25.2
requests==2.31.0

# 数据验证
pydantic==2.5.0
pydantic-settings==2.1.0

# 安全认证
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# 工具库
python-dotenv==1.0.0
python-multipart==0.0.6
schedule==1.2.0

# 数据处理
pandas==2.1.4
numpy==1.24.4

# 日志和监控
loguru==0.7.2

# 异步支持
asyncio-mqtt==0.16.1
aiohttp==3.9.1

# 配置管理
pyyaml==6.0.1
